requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
composio
composio-openai-agents
openai-agents